backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
stripe==7.8.0
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.1
python-jose==3.3.0
cryptography==41.0.7
passlib==1.7.4
cors==1.0.1
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0
alembic==1.13.0
